+++ lib/sapoci/connect/middleware/background_search.rb
@@ -1,0 +1,22 @@
+module SAPOCI
+  module Connect
+    module Middleware
+      class BackgroundSearch < Faraday::Response::Middleware
+        def initialize(app, options = {})
+          @options = options
+          super(app)
+        end
+
+        def call(env)
+          @app.call(env).on_complete do |resp|
+            case env[:status]
+            when 200
+              env[:body] = SAPOCI::Document.from_html(env[:body])
+            else
+            end
+          end
+        end
+      end
+    end
+  end
+end