+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+*2011-08-30 (0.1.1)*
+
+* Simplify API by separating connection initialization (Faraday) and OCI
+  handling via Faraday middleware.
+
 *2011-08-29 (0.1.0)*
 
 * Initial version, supporting OCI background search only.